+++ DTCD-WorkspaceSystem/src/WorkspaceSystem.js
@@ -1,3 +1,5 @@
+import { GridStack } from 'gridstack';
+
 import './styles/panel.scss';
 import './styles/modal.scss';
 import 'gridstack/dist/gridstack.min.css';
@@ -12,14 +14,14 @@ import {
   NotificationSystemAdapter,
 } from './../../DTCD-SDK/index';
 
-import { version } from './../package.json';
-import { GridStack } from 'gridstack';
 
-import gridstackOptions from './utils/gridstackOptions';
-import TabsSwitcher from './TabsSwitcher';
+import { version } from './../package.json';
 import utf8_to_base64 from './libs/utf8tobase64';
-import TabsPanelComponent from './TabsPanelComponent';
+import gridstackOptions from './utils/gridstackOptions';
 import createWidgetErrorMessage from './utils/createWidgetErrorMessage';
+import TabsSwitcher from './TabsSwitcher';
+import TabsPanelComponent from './TabsPanelComponent';
+import gridStackItemHtml from './templates/gridStackItem.html';
 
 const replaces = {
   LiveDashPanel_SimpleMath: 'LiveDashPanel',
@@ -85,6 +87,7 @@ export class WorkspaceSystem extends SystemPlugin {
       'WorkspaceTabSelectedProgrammly',
       'WorkspaceTabClicked',
       'WorkspaceTitleLoaded',
+      'WorkspaceEditModeChanged',
     ]);
     this.#interactionSystem = new InteractionSystemAdapter('0.4.0');
     this.#logSystem = new LogSystemAdapter('0.5.0', this.#guid, 'WorkspaceSystem');
@@ -131,6 +134,11 @@ export class WorkspaceSystem extends SystemPlugin {
       });
     }
 
+    const editModeSwitchAttrs = {
+      label: 'Редактировать рабочий стол',
+    };
+    if (this.#editMode) editModeSwitchAttrs.checked = true;
+
     return {
       fields: [
         {
@@ -156,12 +164,12 @@ export class WorkspaceSystem extends SystemPlugin {
         {
           component: 'switch',
           propName: 'editMode',
-          attrs: {
-            label: 'Редактировать рабочий стол',
-          },
+          attrs: editModeSwitchAttrs,
           handler: {
             event: 'input',
-            callback: this.changeMode.bind(this),
+            callback: (event) => {
+              this.changeMode(event.target.checked);
+            },
           },
         },
         {
@@ -249,7 +257,7 @@ export class WorkspaceSystem extends SystemPlugin {
     const { title, column } = config;
     this.#currentTitle = title;
     this.setColumn(column);
-    this.saveConfiguration();
+    // this.saveConfiguration();
   }
 
   async init() {
@@ -309,7 +317,14 @@ export class WorkspaceSystem extends SystemPlugin {
     this.#tabsSwitcherInstance.htmlElement.addEventListener('tab-delete', this.#handleTabsSwitcherDelete);
     this.#tabsSwitcherInstance.htmlElement.addEventListener('tab-add', this.#handleTabsSwitcherAdd);
     this.#tabsSwitcherInstance.htmlElement.addEventListener('tab-copy', this.#handleTabsSwitcherCopy);
-    this.#eventSystem.subscribe(this.#guid, 'WorkspaceTitleLoaded', 'HeaderPanel_top', 'showTitle');
+    
+    this.#eventSystem.subscribe({
+      eventGUID: this.#guid,
+      eventName: 'WorkspaceTitleLoaded',
+      actionGUID: 'HeaderPanel_top',
+      actionName: 'showTitle',
+      subsctiptionType: 'system',
+    });
 
     const workspaceID = history.state.workspaceID;
     this.setConfiguration(workspaceID).then(() => {
@@ -867,13 +882,13 @@ export class WorkspaceSystem extends SystemPlugin {
     }
   }
 
-  changeMode() {
-    this.#editMode = !this.#editMode;
+  changeMode(doEdit) {
+    doEdit ?? (doEdit = !this.#editMode);
+    this.#editMode = Boolean(doEdit);
 
     const panelBorder = document.querySelectorAll('.grid-stack-item');
     panelBorder.forEach(gridCell => {
-      if (this.#editMode) gridCell.classList.add('grid-stack-item_editing');
-      else gridCell.classList.remove('grid-stack-item_editing');
+      gridCell.classList[this.#editMode ? 'add' : 'remove']('grid-stack-item_editing');
     });
 
     const margin = this.#editMode ? '2px' : '0px';
@@ -888,6 +903,9 @@ export class WorkspaceSystem extends SystemPlugin {
       this.#vueComponent.editMode = this.#editMode;
     }
 
+    this.#eventSystem.publishEvent('WorkspaceEditModeChanged', {
+      editMode: this.#editMode,
+    });
     this.#logSystem.info(`Workspace edit mode turned ${this.#editMode ? 'on' : 'off'}`);
   }
 
@@ -1549,42 +1567,19 @@ export class WorkspaceSystem extends SystemPlugin {
       targetGrid ??= this.#activeGrid;
     }
 
+    const gridStackItemEl = document.createElement('div');
+          gridStackItemEl.className = `grid-stack-item${this.#editMode ? ' grid-stack-item_editing' : ''}`;
+          gridStackItemEl.innerHTML = gridStackItemHtml;
+
+    if (empty) {
+      gridStackItemEl.setAttribute('data-empty-item', '');
+    } else {
+      gridStackItemEl.querySelector('.VisualizationContainer-js')
+                  .setAttribute('id', `panel-${guid}`);
+    }
+
     const widget = targetGrid.addWidget(
-      `
-      <div
-        class="grid-stack-item ${this.#editMode ? 'grid-stack-item_editing' : ''}"
-        ${empty ? ' data-empty-item' : ''}
-      >
-        <div class="grid-stack-item-content">
-          <div class="handle-drag-of-panel gridstack-panel-header">
-            <button
-              class="fix-panel-button"
-              type="button"
-              title="Зафиксировать панель"
-            >
-              <span class="FontIcon name_location size_lg"></span>
-            </button>
-            <button
-              class="close-panel-button"
-              type="button"
-              title="Удалить панель"
-            >
-              <span class="FontIcon name_closeBig size_lg"></span>
-            </button>
-            <button
-              class="drag-panel-button"
-              type="button"
-              title="Переместить панель"
-            >
-              <span class="FontIcon name_move size_lg"></span>
-            </button>
-          </div>
-          <div class="gridstack-content-container">
-            ${empty ? '' : `<div id="panel-${guid}"></div>`}
-          </div>
-        </div>
-      </div>
-      `,
+      gridStackItemEl,
       {
         x,
         y,